# feat: stop sigprof while calling native