--- a/decoder/primitives.go
+++ b/decoder/primitives.go
@@ -30,7 +30,10 @@ func decodeTypedPointer(s string, i int, vt *rt.GoType, vp unsafe.Pointer, sb *_
         return 0, err
     } else {
         rt.MoreStack(_FP_size + _VD_size + native.MaxFrameSize)
-        return fn(s, i, vp, sb, fv, "", nil)
+        rt.StopProf()
+        ret, err := fn(s, i, vp, sb, fv, "", nil)
+        rt.StartProf()
+        return ret, err
     }
 }