--- a/internal/rt/gcwb.go
+++ b/internal/rt/gcwb.go
@@ -74,3 +74,45 @@ func GcwbAddr() uintptr {
         return uintptr(fp) + off
     }
 }
+
+// WARN: must be aligned with runtime.Prof
+// type Prof struct {
+//     signalLock uint32
+// 	hz int32
+// }
+
+var (
+    // // go:linkname runtimeProf runtime.prof
+    // runtimeProf Prof
+
+    // count of native-C calls
+    yieldCount uint32
+
+    // previous value of runtimeProf.hz
+    oldHz int32
+)
+
+//go:nosplit
+func MoreStack(size uintptr)
+
+func StopProf()
+
+// func StopProf() {
+//     atomic.AddUint32(&yieldCount, 1)
+//     if runtimeProf.hz != 0 {
+//         oldHz = runtimeProf.hz
+//         runtimeProf.hz = 0
+//     }
+// }
+
+func StartProf()
+
+// func StartProf() {
+//     atomic.AddUint32(&yieldCount, ^uint32(0))
+//     if yieldCount == 0 && runtimeProf.hz == 0 {
+//         if oldHz == 0 {
+//             oldHz = 100
+//         }
+//         runtimeProf.hz = oldHz
+//     }
+// }