… internal/rt/asm.s