## Backgroud
- When go received a SIGPROF signal while it is running on sonic's native functions, it will crash duet to missing stack information. At present, sonic doesn't support add stack map for native functions (on plan), thus a work-around way is to set  runtime.prof.hz = 0, thus `sigprof()` handler won't call `gentraceback()` which cause crash. After we support make stackmap in future, this behavior will be removed

## Benchmark
- Get
```
name                      old time/op    new time/op    delta
GetOne_Parallel_Sonic-16    3.54µs ±39%    4.85µs ±27%  +36.92%  (p=0.004 n=10+9)
GetOne_Sonic-16             14.9µs ±20%    12.7µs ±14%  -14.79%  (p=0.001 n=10+9)

name                      old speed      new speed      delta
GetOne_Parallel_Sonic-16  3.84GB/s ±43%  2.76GB/s ±34%  -28.22%  (p=0.004 n=10+9)
GetOne_Sonic-16            879MB/s ±18%  1009MB/s ±16%  +14.71%  (p=0.004 n=10+10)

name                      old alloc/op   new alloc/op   delta
GetOne_Sonic-16              24.0B ± 0%     24.0B ± 0%     ~     (all equal)
GetOne_Parallel_Sonic-16     24.6B ± 2%     24.6B ± 2%     ~     (p=1.000 n=10+10)

name                      old allocs/op  new allocs/op  delta
GetOne_Sonic-16               1.00 ± 0%      1.00 ± 0%     ~     (all equal)
GetOne_Parallel_Sonic-16      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
```
- flame graph before this PR
![image](https://user-images.githubusercontent.com/22593627/209899108-6ea881dc-409c-4c01-a1ca-a3721c47e0ce.png)

- flame graph after this PR
![image](https://user-images.githubusercontent.com/22593627/209899021-ae4582d7-b410-4f50-b6c0-192b687295ca.png)